requirements.txt
uvicorn~=0.30.1
fastapi~=0.111.1
pydantic~=2.7.4
passlib~=1.7.4
pymongo~=4.8.0
argon2-cffi~=23.1.0
python-decouple

pandas~=2.2.2
gdown~=5.2.0
pyarrow
